Connections package: refactor, extract classes
