+++ megamek/src/megamek/common/net/connections/INetworkPacket.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2023 - The MegaMek Team. All Rights Reserved.
+ *
+ * This file is part of MegaMek.
+ *
+ * MegaMek is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * MegaMek is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with MegaMek. If not, see <http://www.gnu.org/licenses/>.
+ */
+package megamek.common.net.connections;
+
+/** This is an interface for a connection layer data packet. */
+interface INetworkPacket {
+
+    /** @return The data marshalling type. */
+    int getMarshallingType();
+
+    /** @return The packet data. */
+    byte[] getData();
+
+    /** @return True if the data is compressed. */
+    boolean isCompressed();
+}
